Corrige consulta com parametros e adiciona specification de characters

+++ src/main/java/com/api/marvel/model/DataContainer.java
@@ -11,11 +11,11 @@ public class DataContainer<T> {
 
 	private int offset;
 	private int limit;
-	private int total;
+	private long total;
 	private int count;
 	private List<T> results;
 
-	public DataContainer(Integer offset, Integer limit, int total, int count, List<T> results) {
+	public DataContainer(Integer offset, Integer limit, long total, int count, List<T> results) {
 		this.offset = offset == null ? MarvelAPIConstantes.DEFAULT_PAGINATION_OFFSET : offset;
 		this.limit = limit == null ? MarvelAPIConstantes.DEFAULT_PAGINATION_LIMIT : limit;
 		this.total = total;
